pyproject.toml
[project]
name = "fletube"
version = "1.0.0"
description = "Aplicativo desktop Windows para download e gerenciamento de vídeos"
authors = [
    { name = "Alison Santos", email = "Alisondev77@hotmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flet==0.28.3",
    "requests",
    "python-dotenv",
    "yt-dlp",
    "pytz",
    "cryptography",
    "supabase",
    "loguru",
]

[tool.flet]
name = "fletube"
org = "com.aldev"
product = "Fletube"
company = "Alison"
copyright = "Copyright (C) 2025 by Alison"

[tool.flet.windows]
icon = "assets/icon.png"

[tool.flet.assets]
src = "assets"

[tool.poetry]
package-mode = false